Survive reloads / being disabling before other plugins
When the server disables our plugin, it will also unload all our
classes. This clears the static cache map in EventProxy and throws away
the ClassLoader. Sadly our plugin instance still persists with the
original ClassLoader and is picked up by the injector. It then thinks it
has never encountered that event and faithfully creates a proxy class -
which collides with the leftover classes in the plugin's original
ClassLoader.

To circumvent that we now check if EventTracer is non-null (which it
should always be) and *enabled*. If it is not enabled, it was unloaded
before the plugin causing the current event and wrapping is skipped.